backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.3",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "supabase>=2.24.0",
    "tavily-python>=0.7.13",
    "uvicorn>=0.38.0",
]
